Set default send/recv size to 1M
If MaxWrite and MaxRead are larger than the recv/send buffer sizes, you
will get mysterious and inexplicable errors under krb5i and krb5p.  This
bumps the send/recv size to match the default of the Linux kernel.

This is not an ideal solution.  If you make the maxread/maxwrite
bigger, it will break again and you'll have to explicitly set the
MaxRPCSendBufferSize and MaxRPCRecvBufferSize variables in the
configuration file.

The long term solution is to get rid of these parameters, make Maxread
and Maxwrite global options, and use them instead.

Signed-off-by: Adam C. Emerson <aemerson@linuxbox.com>
Signed-off-by: Matt Benjamin <matt@linuxbox.com>