Add some junction support to cache_inode
The big addition is that cache_inode_getattr must be able to cross a
junction at the request of the call back routine. The way this will work
is that the call back routine will return CACHE_INODE_CROSS_JUNCTION
which will signal cache_inode_getattr to cross the junction and make a
new callback with the node on the other side of the junction.

The callback routine is responsible for any tracking it must do to
know it has been called back with the node across the junction. The
callback routine is also responsible for handling any export permission
considerations of the junction crossing.

This capability is simply a helper and is to allow enforcing of proper
use of the attribute lock. We don't want to acquire the lock for the
node accross the junction while holding the lock for the junction node.
This can especially be an issue with a mid-filesystem junction where
there the junction node and the root node are one and the same.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>