Add new junction support to NFS v4 READDIR
This is all handled inside the callback. When called back on a junction
node, we cross the junction, doing export permission checking and
generating new creds.

If the new export is not available to the particular client because of
export permissions, the entry is completely hidden from that client.

If the client is not using the right authentication flavor, but is just
requesting FATTR4_MOUNTED_ON_FILEID, then we allow the operation to
proceed (AIX clients depend on this).

Otherwise, if the authentication is not right, we report the
NFS4ERR_WRONGSEC in the FATTR4_RDATTR_ERROR field (if present, otherwise
the whole READDIR operation must fail).

Otherwise, the junction is crossed, and we signal cache_inode_getattr to
callback with the node across the junction. We process it.

In all cases, the export and credentials are restored.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>