Final fix (hopefully) of the VFS handle size
Matt's (or was it Adam) last fix for the VFS handle size has
missed a spot and might have broken FBSD. (Actually FBSD was not
broken by chance)

There was already a vfs_sizeof_handle() that was suppose to
abstract the different systems. But it was never used except in
one place. The new vfs_file_handle_size() was now used elsewhere
but was not taking into consideration the FBSD side.

Unite the two into an OS specific member, Also move VFS_FILE_HANDLE_MIN
to be os specific.

This patch fixes an old standing bug with recalls at VFS pnfs side.
(We had a fix in Panasas based on dev42 for a while, now also fixed
for mainline)

Viji do not apply, this exposes a bad recall bug at Ganesha which
is now ignored because of this bug. The other one is worked on.
In the lab we should only work with the two fixes applied together

(Adam Matt please review and commit through your tree, with sign-off)
Signed-off-by: Boaz Harrosh <bharrosh@panasas.com>
Signed-off-by: Adam C. Emerson <aemerson@linuxbox.com>