Add check and refcount to prevent junctions and export roots from unlink
Any cache inode that is an active junction must be protected from unlink
as must any node that is an export root. Since an inode may be the root
of more than one export (with a different tag and/or Pseudo path), we
must refcount export root nodes rather than flagging the node as a root.

Note that this does not prevent non-Ganesha tasks from unlinking these
nodes. That is a task for another day.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>