RPM: new way of managing RPM.
The rpm spec file is no more generated on the fly.
It becomes a configure_file managed by CMake (name of the source tarball has to be set in it).
The spec file is now part of the generated tarball, making it possible to run "rpmbuild -ta <tgz file>"

By default, the specfile will generate libs only for FSAL_VFS and FSAL_PROXY. More fsals can be compiled
and packaged by using the "--with" option in rpmbuild.
For example: rpmbuild --with lustre -ta <tgz file>

This option works for rpmbuild --rebuild : rpmbuild --with xfs --rebuild <src.rpm file>

Signed-off-by: Philippe DENIEL <philippe.deniel@cea.fr>