This patch adds option to the mount.9P helper:
It adds option "cache=" to make it possible to pass "-o cache=mmap" to 9p mount.
This mount option is a requirement for fixed 9p's mmap bug.
The cache=mmap mount option for 9p comes for kernel 3.14

It adds option "context=<some selinux context>" to make it possible to set the default selinux content for a 9p mount point.

Signed-off-by: Philippe DENIEL <philippe.deniel@cea.fr>