Remove need for FSAL_UP to utilize the export and object belongs to
FSAL_VFS stores up_ops in it's private vfs_fsal_obj_handle. This may not
be ideal, but is a solution.

FSAL_GPFS already has a pointer to the up_ops for each FSAL_UP thread.

notify_device is now passed the exportid. Ultimately exportid will not
be part of device id, at which point presumably the uint64_t exportid
parameter will be renamed, but this step allows us to disassociate
export from fsal_obj_handle.

For FSAL_CEPH we also keep the up_ops in the private handle.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>