Improve config capabilities
Add CONFIG_LIST_BITS and CONFIG_ENUM_BITS: These config items allow a
list or enum to occupy a range of bits in a uint32_t, there can be other
config items that occupy other bits (CONF_ITEM_BOOLBIT works just fine
with this without any change). These items have a mask that defines the
set of bits belonging to this option. These items overload CONFIG_LIST
and CONFIG_ENUM (with those having a mask of UINT32_MAX).

Add some debug.

For CONFIG_ENUM, the default and comnfigured values must be set with |=
instead of = for the bit fields to work right.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>