Enable config of export perms for EXPORT_DEFAULTS, EXPORT, and CLIENT
Along the way, deprecate some unused options and unify so all perms
options are set at all levels.

A new config block, EXPORT_DEFAULTS allows setting export permissions
to be used for all exports if not otherwise specified. This utilizes
the new capability of the config parser to track if an option has been
specified or not.

To establish the effective export permissions, each option is
effectively considered in turn (however all bit field options are
actually considered in parallell). If the option has been set in the
applicable CLIENT block, that value will be used, otherwise if set
in the EXPORT block, that value will be used, otherwise if set in
the EXPORT_DEFAULTS block, that value will be used, and finally, if
no configuration block has provided a value, then the code default
value will be used.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>