Add reopen object method
If a file is already open, and we need to open the file with different
open flags, we can use reopen method to avoid closing the file which
would take away all the locks we placed on the file.

This commit introduces reopen method and uses it if an FSAL provides it.

Signed-off-by: Malahal Naineni <malahal@us.ibm.com>