Open files read-only for shared mode locking if FSAL has reopen method.
If a file is already opened in read-only mode, and needs to be opened in
read-write mode, the file needs to be closed and opened again.  Closing
a file loses all kinds of locks placed on the file. For this reason,
even for satisfying a read mode lock (shared lock), we open the file in
read-write mode.

If FSAL has reopen method, we can just open the file in read-only mode.

Signed-off-by: Malahal Naineni <malahal@us.ibm.com>