Use __thread instead of pthread_get_specific in logging code
Convert the thread-specific data to thread local storage,
so we eliminate the expensive pthread_[get|set]specific code.

This also eliminates the need for the emergency_context because
we can static initialize the thread local storage.

Also, trim maximum thread name, and make it static so we aren't
dynamically allocating it.

Signed-off-by: Poornima Gupte <poornima.gupte@in.ibm.com>
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>