Pass cache_entry_t in getattr callback to fix NFS v4 junction crossing
After crossing a junction, the handle returned in NFS v4 READDIR was
that of the junction entry, not that of the root directory of the export.

By passing the cache entry in the callback, we get to pass the proper
entry across the junction instead of passing the opaque data which had
the cache entry.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>