Fix issue if no access allowed to export to allow ENOENT
If no access is granted to a client for an export, the expectation was
that LOOKUP would return ENOENT and READDIR would skip the entry.

There used to be a check at the end of nfs_export_check_access for this
condition, but now it always returns export_perms with options set.

The solution is to test the export_perms.options to make sure some
access is granted. This only affects NFS v4 currently.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>