Don't confuse the BUILDROOT with the final install location
This is a bit tricky.  We must build with the installed files thinking
they are in CMAKE_INSTALL_PREFIX/* but rpmbuild does the install into
a buildroot that is then slurped up into the package as if it is starting
at '/'.  CMAKE_INSTALL_PREFIX is for configure/build time and DESTDIR is
for rpmbuilds %install time.

Signed-off-by: Jim Lieb <jlieb@panasas.com>