Fix FSAL_GPFS lookup to work if run on older GPFS without fsid in getattrs
Older GPFS doesn't pass back the fsid in the attributes. We need
to work around that condition.

Fix gpfs_lookup_path to use fsid from handle not attributes.

Fix GPFSFSAL_getattrs to initialize the fsid to 0.0, make the call into
GPFS, and if the fsid is still 0.0, set the fsid from the fsal_filesystem.

Move the XDEV processing into GPFSFSAL_lookup so that it can check the
fsid in the handle for XDEV crossing, and then pass the correct
fsal_filesystem to GPFSFSAL_getattrs.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>
Acked-by: Venkateswararao Jujjuri (JV) <jvrao@linux.vnet.ibm.com>

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>