Make a more robust vfs_check_handle
Copy code from vfs_create_handle to use to check handles also
for extract_handle. This will do more checking earlier.

It also sets the stage for checking if a handle has a valid
filesystem for the given export. We won't put that bit in until
we clean up some locking and refcounting though.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>
Acked-by: Jim Lieb <jlieb@panasas.com>