Use extended error handling in export manager's DBus 'AddExport' method
Uses detailed errors to decide what happened in the config processing.  This
allows the user to add exports to a base configuration and use that file
for the AddExport method.  It's behavior is as follows:

1. All exports found are processed.

2. All parse/processing errors for all exports are reported.

3. If there is already a configured export of the same id, the
   new version is skipped.  This means you cannot modify an export this way.

4. Any new export is added and will appear in the namespace.

This method does not allow the picking and choosing from the set of (potentially)
new exports.

The AddExport method either returns an error with a error summary or a message
containing the number of exports added.  The user will have to consult the server
log for details (file and line number etc.)

Signed-off-by: Jim Lieb <jlieb@panasas.com>