Fix assertion failure due to errors in facility block
facility->fac_list is initialized with glist_init() in facility_init(),
so it becomes glist_empty rather than glist_null when a "facility block"
fails to materialize due to errors. The assertion that it should be
"glist_null" fails in this case.

facility->fac_list is an embedded list_head, so glist_init() should NOT
be called on it.

We have two kinds of glist_heads. One is a list anchor and other is an
embedded glist_head inside a node (aka glist_member). glist_init()
should only be called on the former, and glist_empty() checks can be
done on it. The latter glist_head type should be initialized with NULL
pointers (no need, if allocating the node with calloc) and glist_null
can be called on it. Don't call glist_empty() on these embedded
glist_heads.

Signed-off-by: Malahal Naineni <malahal@us.ibm.com>