Downgrade open flags when a state is removed or downgraded
Currently if two processes open the file, one in read-only mode and the
other in read-write/write mode, the file is kept open in read-write mode
as long as either of them keep the file open. With this patch, the file
is downgraded to read-only mode when the second process closes the file.

Signed-off-by: Malahal Naineni <malahal@us.ibm.com>
Reviewed-by: Frank S. Filz <ffilzlnx@mindspring.com>
Reviewed-by: Venkateswararao Jujjuri <jvrao@linux.vnet.ibm.com>