Save sle_type to fix use after free in state_export_unlock_all (fixes CID 30405)
This fixes the apparent intent of the code, however, it should be noted that
while this is definitely correct for range locks (sle_type == POSIX_LOCK), it
MAY not be correct for delegations (sle_type == LEASE_LOCK). On the other hand,
state_export_release_nfs4_state doesn't do anything else to clean up'
delegations.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>
Acked-by: Malahal Naineni <malahal@us.ibm.com>