Remake cache_inode_destroyer into disposing of AVL tree instead of entries
Unexporting all exports as part of shutdown flushes the cache inode, so
instead, we will destroy the AVL tree. If the tree is NOT empty, we will
throw an error.

The function is moved into cache_inode_init.c and cache_inode_hash.c to
be adjacent to the initialization code for ease of verifying that destroy
undoes everything that init does.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>
Acked-by: Malahal Naineni <malahal@us.ibm.com>