Initial pass of new device id
This implements the new device id with the first byte being an FSAL_ID.

Rather than the rest of the deviceid being an entirely opaque string of
15 bytes, it has been divided up into a uint8_t, a uint16_t, a uint32_t,
and a uint64_t. The deviceid is put on the wire in host byte order, but
these fields make it a little easier for some of the existing PNFS
implementations.

I have disconnected some bits, and almost certainly no one's PNFS
works now.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>