SHOOK returns: this commits repairs FSAL_LUSTRE (broken in 2.0)
The FSAL_SHOOK was broken since new_api was introduced in 2.0.
This commit repairs this feature by managing "shook restore"
from the FSAL itself.
FSAL_SHOOK still derives from FSAL_LUSTRE.
An option in Cmake (with a cmakedfine in src/include/config.h) is
used to enable/disable this feature.

Signed-off-by: Philippe DENIEL <philippe.deniel@cea.fr>