Add kill_export_junction_entry to handle ESTALE on a junction inode
If we lose a junction inode, we don't remove the export but we must
unmount from the PseudoFS.

This patch also cleans up a bit of unmounting from the PseudoFS stuff
to make sure it doesn't skip stuff if exp_junction_inode is NULL.

Basically we always clean up the PseudoFS an export is mounted on.

Also, if exp_junction_inode is non-NULL, we clean up that.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>