Fix fsal_filesystem indexing issues
If libblkid is not present, we will use the fsid from stat_fs. The minor
and major fields in that are int32_t. Directly converting to uint64_t
entails sign extending a negative number first, which results in values that
don't match as expected when only 32 bits for each field are put into the
wire handle. By casting to unsigned first, the sign extension does not occur.

There was also a minor issue in that fsid_type was not set before inserting a
filesystem into the database if fsid_type was changed. Now that the comparison
function takes fsid_type into account (to handle FSID_MAJOR_64), this must
be changed before insert to the table.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>