Fix errno 34 (ERANGE) with getgrouplist() call
The first getgrouplist call with 0 ngroups returns -1 as the man page
says, but errno was set to 34 (instead of 0) under some environments.
The man page doesn't talk about errno with regards to getgrouplist call.
It did set the ngroups to the expected value!

Removed the errno check with getgrouplist call. The code assumes that
ngroups is set correctly no matter what the errno value is.

Reported-by: John Dorfner <dorfner@us.ibm.com>
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>
Acked-by: Jim Lieb <jlieb@panasas.com>
Reviewed-by: Venkateswararao Jujjuri <jvrao@linux.vnet.ibm.com>
Acked-by: Frank S. Filz <ffilzlnx@mindspring.com>