idmapper collision fix
The threads that lookup by name or id use the read lock. If they don't
find an entry, then they release the read lock, acquire the write lock
and then add the entry. So it is possible that multiple threads may fail
to find an entry at one point and they all try to add. In this case, we
will be trying to insert same name,id mapping. It is also possible that
name got a different id or an id got a different name causing us to find
an existing entry when we are trying to add an entry! So collison
warning log message is incorrect!

Removed the collision warning log message. Re-arranged the code to make
it a bit more readable. Also added in_uidtree boolean field in struct
cache_user to avoid removing a node entry without actually inserting one
before!

Reported-by: John Dorfner <dorfner@us.ibm.com>
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>