Add export option to enable use of readdir results for negative dirent cache
Most current FSALs do not provide the proper cache coherency to support use
of the readdir results to perform negative dirent caching. Further, some
filesystems support implied files, which are not enumerated in readdir
results but can be looked up directly, in such a case, even with proper
invalidation of dirent cache, use of the dirent cache to infer negative
lookup results would prevent looking up one of these implied files once
a readdir had been performed.

This behavior is controled by a new export option, which defaults to false
in support of the bulk of current FSALs and filesystems.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>
Acked-by: Malahal Naineni <malahal@us.ibm.com>
Acked-by: Jim Lieb <lieb@sea-troll.net>
Acked-by: Soumya Koduri <skoduri@redhat.com>
Acked-by: Matt Benjamin <matt@linuxbox.com>