Fix for coverity CID24725
The allocated string buffer is not enough to hold the original string
buffer, but we do copy only a partial string.  This is clearly a false
positive from coverity as it doesn't really look at how the string is
used! Use gsh_strdup as this is more cleaner, and as a bonus avoids
this false positive from coverity!

Signed-off-by: Malahal Naineni <malahal@us.ibm.com>
Reviewed-by: Venkateswararao Jujjuri (JV) <jvrao@linux.vnet.ibm.com>