Comments useless code. This code will be uncommented as code will be added in the config's commit functions.
FSAL_PROXY and FSAL_LUSTRE have block with *_commit functions, to be called during parsing to check configuration item.
Those functions are currently empty are are basically placeholders.
Setting a variable withut using it triggers a compilation error in Maintainer mode, so I commented this.
This piece of code will be uncommented as "flesh" will come in this part of the code.

Change-Id: I266d8000fa689e3f62e2d97b94f5cc9f69c54e9f
Signed-off-by: Philippe DENIEL <philippe.deniel@cea.fr>