Attempt to only add one client to export per entry (fixes BZ 1074481)
The client list was adding two entries for each possible address because
getaddrinfo shows an entry for both UDP and TCP. A simple check is to keep
track of the last address added, and not add it again. This should eliminate
all the duplicates (especially if JUST an ip address is given in which case
the addrinfo will ONLY return those two entries).

Also for showmount -e report the Wildcard and IPv6 addresses not as
unknown.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>
Acked-by: Malahal Naineni <malahal@us.ibm.com>
Acked-by: Soumya Koduri <skoduri@redhat.com>
Acked-by: mmadhusu@redhat.com