Use non-printable character instead of DOT for revoked file handles
Revoked delegation file handles are stored as a file name in the client
recovery directory by prefixing a "." to the base16 encoding of the file
handle. nfs4_clean_old_recov_dir() assumes that anything that starts
with "." inside base recovery directory is a revoked file handle but a
client specific recovery directory may have components that start with
"." as well.

This patch prefixes non-printable character '\x1' to the base16 encoding
of the revoked file handle to create file name. This should not conflict
with client recovery directory as it will have only printable
characters.

Also fixed nfs4_clean_old_recov_dir() to not make recursive calls on
revoked file handle names.  Also, removed an unneeded memset to 0.

Signed-off-by: Malahal Naineni <malahal@us.ibm.com>
Reviewed-by: Frank S. Filz <ffilzlnx@mindspring.com>
Acked-by: Venkateswararao Jujjuri (JV) <jvrao@linux.vnet.ibm.com>