Send sm_notify to affected clients only.
When an IP is released from a server node, send sm_notify to
clients that are connected to that particular ip addr only.
Added a routine to match strings if they are in ipv6 format or ipv4
format. In the testing we found strings to comare are like:
 "10.0.100.231"  and  "::ffff:10.0.100.231"

Ported from 1.5.x: e7e47b9aa1c326d750b8765868389aec63b94ae9

Signed-off-by: Venkateswararao Jujjuri (JV) <jvrao@linux.vnet.ibm.com>
Reviewed-by: Malahal Naineni <malahal@us.ibm.com>
Acked-by: Rong Zeng <rongzeng@us.ibm.com>