Fix claim_posix_filesystems to work with bind mounts
claim_posix_filesystems() uses mount point path to find the file system
image for a given export path. Bind mounts will not be in the list of
file system images as they are actually duplicates. In other words,
mount point of a bind mount will not be in the list, and any export path
from a bind mount path fails to match any file system image.

This patch uses st_dev field from stat system call on the export path to
look for a matching file system image.

Signed-off-by: Malahal Naineni <malahal@us.ibm.com>
Reviewed-by: Venkateswararao Jujjuri (JV) <jvrao@linux.vnet.ibm.com>
Reviewed-by: Frank S. Filz <ffilzlnx@mindspring.com>