FSAL_LUSTRE/Upcalls: this commit implements FSAL Upcalls for FSAL_LUSTRE
This commit implements the basic support of upcalls in FSAL_LUSTRE.
The mechanism is based on Lustre changelogs feature. For scalability
reason, changelog are accessed through the lcap_* API intsead of the
llapi_* (lcap provides changelog caching which leverage loads on the
Lustre MDS). Upcalls are generated by a thread, just the way GPFS does
and uses fsal_up_top function vector.
This commit needs to have lustre support for jobid in changelog records.
This means that LUSTRE 2.7 or higher is required to compile this commit.
The feature is triggered by cmake and/or rpmbuild. By default it is OFF.

Change-Id: I3f4819c99c20902b0ea4691d40d4976ff8962fe4
Signed-off-by: Philippe DENIEL <philippe.deniel@cea.fr>