Introduce return value to get_gsh_export_ref to prevent new refs to stale export
If an export is stale (being unexported for any reason), we should prevent
handing out additional references where possible.

We add a new parameter for those cases for which we need to take a reference
anyway. Also, this paremeter is temporarily used for cases where dealing with
the failure to get a reference is not easy. These cases WILL be fixed in later
patches (but use of this parameter allows us to introduce the feature without
making things any worse). There will be at least one place where this parameter
will still be useful.

If export_release_ok is true, the only possible return is true (success) thus
it is legitimate to ignore the return value.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>