Fix user_date_fmt and time_date_fmt resulting in null time strings bug.
Fix bug: Log using user_time_fmt resulted in null time strings.  This is
because user_date_fmt_ and time_date_fmt are passed as fields to a
structure, and not as global statics.  Use the structure fields, and get
rid of the statics which should no longer be used.

Config file fragment for testing purposes.
	LOG {
	 Format {
	  date_format = none;
	  time_format = user_defined;
	  user_time_format = "%s";
	 }
	}

Signed-off-by: Marcus Watts <mdw@linuxbox.com>