FSAL_GLUSTER: rework getattrs() to use supported/requested ATTR_*
Add ATTR_ACL to GLUSTERFS_SUPPORTED_ATTRIBUTES and
GLUSTERFS_SETTABLE_ATTRIBUTES. These defines are used when a handle is
constructed by calling fs_supported_attrs(). Afterwards, it is possible
for to change the attrlist->mask to request only specific attributes.

While reviewing the usage of ATTR_* and XATTR_* flags, it seems nicer to
use the FSAL_*_MASK() defines that are intended for modifying the mask.

Signed-off-by: Niels de Vos <ndevos@redhat.com>
Acked-by: Soumya Koduri <skoduri@redhat.com>
Ached-by: Daniel Gryniewicz <dang@cohortfs.com>