VFS sub-FSALs and FSAL_PANFS
This adds some infrastructure for sub-FSALs in FSAL_VFS, and adds a
FSAL_PANFS sub-FSAL.

FSAL_VFS already has one sub-FSAL: FSAL_XFS.  It shares most code with
FSAL_VFS, but in a sort of ad-hoc way.  Currently, PNFS in FSAL_VFS is
implemented by PanFS, with code that only works with the PanFS drivers
installed.

This adds some infrastructure to FSAL_VFS (in the form of subfsal.h)
that each sub-FSAL must implement.  It then moves the VFS-specific code
to a sub-FSAL for VFS, and the PanFS-specific code to a sub-FSAL for
PanFS.

FSAL_VFS and FSAL_PANFS have been tested; FSAL_XFS has been built.

Signed-off-by: Daniel Gryniewicz <dang@fprintf.net>
Acked-by: Frank Filz <ffilzlnx@mindspring.com>