Use state_t for NLM share reservations
In doing so, we eliminate the redundant state_nlm_share structure.

A few bugs have been cleared up along the way.

Also, the RFC does not really specify what happens if the client sends a
2nd NLM_SHARE for the same file and same owner. In order to handle the case,
we assume that such an NLM_SHARE updates the state.

Consequently, an NLM_UNSHARE will remove the specified access/deny bits if
they are part of the current share state.

If the client in fact never does this, it should send paired NLM_SHARE and
NLM_UNSHARE.

Change-Id: I4e17b93321a4fb4416039b0ee42f9ebc335e380e
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>