Cleanup lock ordering issues in cache inode
We MUST NOT take the content_lock while holding the qlane lock. By
passing a new flag to cache_inode_close on cleanup we can assure this.

Another lock reversal happened with cache_inode_dec_pin_ref being called
with closefile true. The fix there is to require the content_lock be held
while calling cache_inode_dec_pin_ref if closefile is true.

There is also a possibility of double lock on the qlane lock if
cache_inode_close results in an ESTALE. This could occur on cleanup or
when cache_inode_dec_pin_ref is called with closefile true. An additional
flag indicates to cache inode functions to NOT call cache_inode_kill_entry
to avoid this double lock deadlock.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>