Make sure cleanup of LAYOUT states terminates
Since we push non-LAYOUT states to the end of the list, the list is effectively
infinite. Track the first state in the list. If we hit it again
without having seen any LAYOUT states, then the list only contains
non-LAYOUT states and we can terminate.

If the first item is a stale LAYOUT state, the glist_for_each_safe loop
will continue until we hit a non-stale LAYOUT state or any remaining
stale LAYOUT states are drained.

If the first item is not a stale LAYOUT state, we are going to clean it
up and restart the loop.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>
Acked-by: Malahal Naineni <malahal@us.ibm.com>