Teach scanner/parser to recognize IP addrs and FSIDs
The IPv6 pattern recognizes '::', the unspecified address.  If you
use something like ::FFFF:0.0.0.0, it will not recognize this, the
unspecified v4 address wrapped in a v6 address. You shouldn't be doing this
anyway.  Use '::' for v6 and 0.0.0.0 or (for ganesha only) '*';  The alternative
for this is also ::ffff:0:0 (you can only use '::' once...

It also recognizes FSIDs as major.minor in any base.

Signed-off-by: Jim Lieb <lieb@sea-troll.net>
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>