CMake: enforce gcc as a C compiler
This commit enforces gcc as the C compiler by setting CMAKE_C_COMPILER to gcc and
CMAKE_COMPILER_IS_GNUCXX to TRUE.
This is useful when coverity is used : coverity expects GCC as the compiler (setup
before coverity runs). If something elese (even aliases) is used, coverity won't produce
results to be analyzed.

Change-Id: I3df81af6fdbb13bfd97c9e35c716bf28a317048f
Signed-off-by: Philippe DENIEL <philippe.deniel@cea.fr>