Allow for permission checking for LINK in cache_inode to fix FSAL_VFS
In FSAL_VFS, linkat can not be called with non-root fsuid etc. the way Ganesha
needs to use it. This means that we can NOT rely on the kernel to do permission
checking.

A new fs_supports option is provided so FSAL can indicate if it can handle
permission checking for LINK. This option is set to true for all FSALs except
FSAL_VFS (and it's sub-FSALs: FSAL_XFS, and FSAL_PANFS).

Change-Id: Ie7a3936714c2f6906f9e0a69a8b34916e5e9af06
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>