Send SECINFO response in a preferred order.
We send AUTH_NONE followed by AUTH_UNIX in the SECINFO response. Recent
Linux clients pick AUTH_NONE instead of AUTH_UNIX because of this.

RFC3530 says:
The result will contain an array which represents the security
mechanisms available, with an order corresponding to server's
preferences, the most preferred being first in the array.  The client
is free to pick whatever security mechanism it both desires and
supports, or to pick in the server's preference order the first one
it supports.

Also reverts commit 9bc3a270d168432bdff648864bf525af582edbea as we can
have both AUTH_NONE and AUTH_UNIX flavors with AUTH_UNIX being the first
one in the list.

Change-Id: I997d251278ce61cbff4058b20b506d8f15d2c524
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>
Acked-by: Matt W. Benjamin <matt@cohortfs.com>
Acked-by: Frank S. Filz <ffilzlnx@mindspring.com>