Add renamed object handle to FSAL rename so FSAL_VFS can handle special inodes
FSAL_VFS has issues with block, char, and socket inodes, it is unable (at least
on some older kernel versions) to open and fstat a file descriptor for these,
so we remember the directory and name so we can do an fstatat by name.
Unfortunately a rename messes this up.

By passing the object handle to the rename method, FSAL_VFS is able to fixup
these entries and they continue to be usable. This increases the number of
unit tests that pass, and may keep these inodes more usable in the long run
(there is still a problem if Ganesha restarts or the inode cache is flushed
while a client has a cached handle for one of these objects).

Change-Id: Ia325bc4ac8b242648a08406090171d357c9ea6fe
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>