introduce FSAL_RGW
This change provides a FSAL consuming a new file-based
interface to Ceph's RADOS Gateway service, consumed as a
library.

The FSAL is in a preliminary state, a variety of interfaces are
not implemented (e.g., upcalls).

The change is a concatenation of the following commits:
    * Basic sekleton for RGW FSAL
    * build FSAL_RGW
    * Add rgw file functions
    * Add nfs_handle
    * remove fsal2ceph error function and inode
    * mount command wip

Change-Id: I239bfd051f049feb77c8deff440f7be64513ddba
Signed-off-by: Orit Wasserman <owasserm@redhat.com>
Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>
(cherry picked from commit 674f265cfe92ce58468bbd277227a7d93823c357)