Fix never ending list traversal under some cases
There are few places where we traverse the list of states to drain some
matching states.  We also put the elements at the end of the list while
traversing the list to avoid having unmatched states at the beginning of
the list. It is possible that the list traversal may not end if the list
has only states that are not matched.

This patch adds 'first' state, and if we encounted it again, that means
we did complete traversal without any states that match. This terminates
the list.

Change-Id: I09048b7451faf43e2e417b883f47dcaec7ffb3b3
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>